rename input
